# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AI-native headless workflow automation platform. DAG execution, MCP-native, queue-based.",
        "title": "aflow API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/workflows": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all workflows in the workspace.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "List workflows",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new workflow with an initial version (definition required).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Create workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Workflow definition",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.createWorkflowReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.createWorkflowResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    }
                }
            }
        },
        "/v1/workflows/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single workflow by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Get workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Workflow UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.workflowResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates name and description. Does not change the definition.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Update workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Workflow UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.updateWorkflowReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.workflowResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently deletes a workflow and all its versions and executions.",
                "tags": [
                    "Workflows"
                ],
                "summary": "Delete workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Workflow UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    }
                }
            }
        },
        "/v1/workflows/{id}/deactivate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sets active=false. Stops cron chains and rejects webhook triggers.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Deactivate workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Workflow UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.workflowResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    }
                }
            }
        },
...
        "/v1/workflows/{id}/publish": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validates the definition, marks the latest version published, sets active=true, and activates its trigger (cron or webhook).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Publish workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Workflow UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.publishWorkflowResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    }
                }
            }
        },
        "/v1/workflows/{id}/versions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all immutable versions of a workflow, newest first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "List workflow versions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Workflow UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_api_handlers.ProblemDetail": {
            "description": "RFC 7807 error envelope returned on all 4xx/5xx responses.",
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Workflow abc was not found"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                },
                "title": {
                    "type": "string",
                    "example": "Not Found"
                },
                "type": {
                    "type": "string",
                    "example": "https://aflow.dev/errors/not-found"
                }
            }
        },
...
        "internal_api_handlers.createWorkflowReq": {
            "type": "object",
            "properties": {
                "definition": {
                    "type": "object"
                },
                "description": {
                    "type": "string",
                    "example": "Processes incoming emails"
                },
                "name": {
                    "type": "string",
                    "example": "Email Processor"
                }
            }
        },
        "internal_api_handlers.createWorkflowResp": {
            "type": "object",
            "properties": {
                "version": {
                    "$ref": "#/definitions/internal_api_handlers.versionResp"
                },
                "workflow": {
                    "$ref": "#/definitions/internal_api_handlers.workflowResp"
                }
            }
        },
...
        "internal_api_handlers.publishWorkflowResp": {
            "type": "object",
            "properties": {
                "version": {
                    "$ref": "#/definitions/internal_api_handlers.versionResp"
                },
                "webhook_url": {
                    "type": "string",
                    "example": "/webhooks/uuid?secret=whsec_..."
                },
                "workflow": {
                    "$ref": "#/definitions/internal_api_handlers.workflowResp"
                }
            }
        },
...
        "internal_api_handlers.updateWorkflowReq": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Updated description"
                },
                "name": {
                    "type": "string",
                    "example": "Email Processor v2"
                }
            }
        },
        "internal_api_handlers.versionResp": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "definition": {
                    "type": "object"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "published": {
                    "type": "boolean",
                    "example": true
                },
                "version": {
                    "type": "integer",
                    "example": 1
                },
                "workflow_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "internal_api_handlers.workflowResp": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "latest_version": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Email Processor"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "workspace_id": {
                    "type": "string",
                    "example": "ws_prod"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT (HS256, claim: workspace_id) or aflow_ API key. Set AFLOW_AUTH_DISABLED=true to skip in dev.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}